Revert "chore: Better organize component folders"
This reverts commit e227e605a980f6436c16d3347c26c43cd017e2bb.
