+++ components/Header.js
@@ -2,10 +2,10 @@ import React from 'react'
 import Image from 'next/image';
 import { useRouter } from 'next/router'
 
-import AddTokenButton from '../buttons/AddTokenButton';
+import AddTokenButton from './AddTokenButton';
 
-import kleverLogo from '../../public/assets/logo.svg'
-import shootingStar from '../../public/assets/shooting-star.svg'
+import kleverLogo from '../public/assets/logo.svg'
+import shootingStar from '../public/assets/shooting-star.svg'
 import Link from 'next/link';
 
 function Header() {
@@ -19,11 +19,11 @@ function Header() {
         </Link>
 
         {/* Image, Text and Button */}
-        <div className='flex justify-center items-center mt-2'>
-          <div className='flex justify-between items-center w-80 sm:w-3/4 max-w-2xl'>
-          <div className='flex justify-start items-center mt-2 w-96'>
+        <div className='flex justify-center items-center p-2'>
+          <div className='flex justify-between items-center p-2 w-80 sm:w-3/4 max-w-2xl'>
+          <div className='flex justify-start items-center p-2 w-96'>
             {/* Image and Text */}
-            <div className='flex justify-center items-center mt-2'>
+            <div className='flex justify-center items-center p-2'>
               {/* Image */}
               <div>
                 <Image src={shootingStar} alt="Shooting star" className='w-10 sm:w-14'/>
@@ -36,7 +36,7 @@ function Header() {
           </div>
 
           {/* Add Button */}
-          <div className='flex justify-end'>
+          <div className='flex justify-end sm:ml-20'>
             { router.pathname === '/' && <AddTokenButton/>}
           </div>
         </div>
