+++ components/MainContainer.js
@@ -1,4 +1,4 @@
-import Header from "./sections/Header"
+import Header from "./Header"
 
 export default function MainContainer({ children }) {
   return (
